♻️ Moved organization admin resolution out of the Organization model
getReplyEmails, getFullAdmins, getFinanceAdmins, getAdminToEmails,
adminsToRecipients, getAdminRecipients, getFinanceAdminRecipients and
getInvoicingToEmail move to OrganizationAdminService in the api app. The
moved code is byte-for-byte identical apart from taking the organization
as a parameter instead of `this`.

getAdmins stays on the model: it is a plain query and has a caller that
only wants the users.

getInvoicingToEmails (plural) is deleted. It has no callers anywhere in
the repo and was superseded by getInvoicingToEmail when invoice XML was
added.

This clears the last three Platform.getSharedStruct() reads from
@stamhoofd/models. The package no longer resolves the platform anywhere
outside the Platform model itself, which completes phase 2.
