+++ backend/shared/models/src/models/Organization.test.ts
@@ -1,8 +1,6 @@
 import { Database } from '@simonbackx/simple-database';
-import { AccessRight, PermissionLevel, PermissionRoleDetailed, Permissions } from '@stamhoofd/structures';
 import { EventFactory } from '../factories/EventFactory.js';
 import { OrganizationFactory } from '../factories/OrganizationFactory.js';
-import { UserFactory } from '../factories/UserFactory.js';
 import { Organization } from './Organization.js';
 import { TestUtils } from '@stamhoofd/test-utils';
 
@@ -117,49 +115,5 @@ describe('Organization.updateFutureEventsForOrganizations', () => {
         await expect(Organization.updateFutureEventsForOrganizations([])).resolves.toBeUndefined();
 
         expect(await getHasFutureEvents(organization)).toBe(true);
-    });
-});
-
-describe('Organization admin recipients', () => {
-    async function createOrganizationWithAdmins() {
-        const financeRole = PermissionRoleDetailed.create({
-            name: 'financial director',
-            accessRights: [AccessRight.OrganizationFinanceDirector],
-        });
-
-        const organization = await new OrganizationFactory({ roles: [financeRole] }).create();
-
-        const fullAdmin = await new UserFactory({
-            organization,
-            email: 'full-admin@example.com',
-            permissions: Permissions.create({ level: PermissionLevel.Full }),
-        }).create();
-
-        const financeAdmin = await new UserFactory({
-            organization,
-            email: 'finance-admin@example.com',
-            permissions: Permissions.create({
-                level: PermissionLevel.None,
-                roles: [financeRole],
-            }),
-        }).create();
-
-        return { organization, fullAdmin, financeAdmin };
-    }
-
-    test('The finance admin recipients include a financial director without full access', async () => {
-        const { organization, fullAdmin, financeAdmin } = await createOrganizationWithAdmins();
-
-        const emails = (await organization.getFinanceAdminRecipients()).map(r => r.email);
-
-        expect(emails).toIncludeSameMembers([fullAdmin.email, financeAdmin.email]);
-    });
-
-    test('The regular admin recipients only include full admins', async () => {
-        const { organization, fullAdmin } = await createOrganizationWithAdmins();
-
-        const emails = (await organization.getAdminRecipients()).map(r => r.email);
-
-        expect(emails).toIncludeSameMembers([fullAdmin.email]);
     });
 });
