+++ backend/shared/models/src/models/Organization.ts
@@ -352,158 +352,10 @@ export class Organization extends QueryableModel {
         });
     }
 
-    /**
-     * E-mail address when we receive replies for organization@stamhoofd.email.
-     * Note that this sould be private because it can contain personal e-mail addresses if the organization is not configured correctly
-     */
-    async getReplyEmails(): Promise<EmailInterfaceRecipient[]> {
-        const sender: OrganizationEmail | undefined = this.privateMeta.emails.find(e => e.default) ?? this.privateMeta.emails[0];
-
-        if (sender) {
-            return [
-                {
-                    name: sender.name,
-                    email: sender.email,
-                },
-            ];
-        }
-
-        return await this.getAdminToEmails();
-    }
-
     async getAdmins() {
         // Circular reference fix
         const User = (await import('./User.js')).User;
         return await User.getAdmins(this.id, { verified: true });
-    }
-
-    /**
-     * These email addresess are private
-     */
-    async getFullAdmins() {
-        const admins = await this.getAdmins();
-        const platform = await Platform.getSharedStruct();
-
-        // Only full access
-        return admins.filter(a => a.permissions && a.permissions.forOrganization(this, platform, { inheritFromPlatform: false })?.hasFullAccess());
-    }
-
-    /**
-     * These email addresess are private
-     */
-    async getFinanceAdmins() {
-        const admins = await this.getAdmins();
-        const platform = await Platform.getSharedStruct();
-        const filtered = admins.filter(a => a.permissions && (a.permissions.forOrganization(this, platform, { inheritFromPlatform: false })?.hasFullAccess() || a.permissions.forOrganization(this, platform, { inheritFromPlatform: false })?.hasAccessRight(AccessRight.OrganizationFinanceDirector)));
-
-        // Only full access
-        return filtered;
-    }
-
-    /**
-     * These email addresess are private
-     */
-    async getAdminToEmails(): Promise<EmailInterfaceRecipient[]> {
-        const filtered = await this.getFullAdmins();
-
-        if (STAMHOOFD.environment === 'production') {
-            if (filtered.length > 1) {
-                // remove stamhoofd email addresses
-                const f = filtered.flatMap(f => f.getEmailTo()).filter(e => !e.email.endsWith('@stamhoofd.be') && !e.email.endsWith('@stamhoofd.nl'));
-                if (f.length > 0) {
-                    return f;
-                }
-            }
-        }
-
-        return filtered.flatMap(f => f.getEmailTo());
-    }
-
-    adminsToRecipients(admins: User[]) {
-        return admins.flatMap((f) => {
-            return Recipient.create({
-                firstName: f.firstName,
-                lastName: f.lastName,
-                email: f.email,
-                replacements: [],
-            });
-        });
-    }
-
-    /**
-     * These email addresess are private
-     */
-    async getAdminRecipients(): Promise<Recipient[]> {
-        const filtered = await this.getFullAdmins();
-        return this.adminsToRecipients(filtered);
-    }
-
-    /**
-     * These email addresess are private
-     */
-    async getFinanceAdminRecipients(): Promise<Recipient[]> {
-        const filtered = await this.getFinanceAdmins();
-        return this.adminsToRecipients(filtered);
-    }
-
-    /**
-     * These email addresess are private
-     */
-    async getInvoicingToEmails() {
-        // Circular reference fix
-        const filtered = await this.getFinanceAdmins();
-
-        if (filtered.length > 0) {
-            return filtered.flatMap(f => f.getEmailTo()).map((recipient) => {
-                if (!recipient.name) {
-                    return recipient.email;
-                }
-                const cleanedName = Formatter.emailSenderName(recipient.name);
-                if (cleanedName.length < 2) {
-                    return recipient.email;
-                }
-                return '"' + cleanedName + '" <' + recipient.email + '>';
-            }).join(', ');
-        }
-
-        return undefined;
-    }
-
-    /**
-     * Returns one email for invoices. since in ubl we can only add one address.
-     * We choose the oldest user that was active in the last 3 months (otherwise the oldest user if noone was active)
-     */
-    async getInvoicingToEmail(): Promise<string | undefined> {
-        // Circular reference fix
-        const admins = await this.getAdmins();
-
-        const tokens = await Token.select().where('userId', admins.map(a => a.id)).fetch();
-
-        // Sort by admins that were active in the last 3 months, then creation date
-        const cutoffDate = new Date(Date.now() - 1000 * 60 * 60 * 24 * 31 * 3);
-        admins.sort((a, b) => {
-            const aTokens = tokens.filter(t => t.userId === a.id);
-            const bTokens = tokens.filter(t => t.userId === b.id);
-            const aActive = !!aTokens.find(t => t.updatedAt > cutoffDate);
-            const bActive = !!bTokens.find(t => t.updatedAt > cutoffDate);
-            return Sorter.stack(
-                Sorter.byBooleanValue(aActive, bActive),
-                Sorter.byDateValue(b.createdAt, a.createdAt),
-            );
-        });
-
-        const platform = await Platform.getSharedStruct();
-        const filtered = admins.filter(a => a.verified && a.permissions && !a.email.endsWith('@stamhoofd.be') && (a.permissions.forOrganization(this, platform, { inheritFromPlatform: false })?.hasFullAccess() || a.permissions.forOrganization(this, platform, { inheritFromPlatform: false })?.hasAccessRight(AccessRight.OrganizationFinanceDirector)));
-
-        if (filtered.length > 0) {
-            return filtered.map(f => f.email)[0];
-        }
-        const filtered2 = admins.filter(a => a.verified && a.permissions && (a.permissions.forOrganization(this, platform, { inheritFromPlatform: false })?.hasFullAccess() || a.permissions.forOrganization(this, platform, { inheritFromPlatform: false })?.hasAccessRight(AccessRight.OrganizationFinanceDirector)));
-
-        if (filtered2.length > 0) {
-            return filtered2.map(f => f.email)[0];
-        }
-        return undefined;
     }
 
     /**
